Actor model updated.

+++ src/main/scala/de/kp/shopify/insight/model/InsightForecast.scala
@@ -80,7 +80,7 @@ case class InsightForecast(
   state:String,
 
   @JsonProperty("amount")
-  amount:Float,
+  amount:Double,
 
   @JsonProperty("time")
   time:Long,
